+++ sdk/lib/crt/wine/undname.c
@@ -234,7 +234,7 @@ static char* str_array_get_ref(struct array* cref, unsigned idx)
 static char* str_printf(struct parsed_symbol* sym, const char* format, ...)
 {
     va_list      args;
-    unsigned int len = 1, i, sz;
+    size_t len = 1, i, sz;
     char*        tmp;
     char*        p;
     char*        t;
